+++ .vscode/settings.json
@@ -10,7 +10,7 @@
   // "rust-analyzer.check.workspace": false,
   // "rust-analyzer.check.features": "all",
   // "rust-analyzer.cargo.buildScripts.rebuildOnSave": false,
-  // "rust-analyzer.check.workspace": false,
+  "rust-analyzer.check.workspace": false,
   "rust-analyzer.cargo.features": "all",
   "rust-analyzer.check.features": "all",
   // "rust-analyzer.check.allTargets": true,
